[Checkpoint] Test file for checkpoint_wrapper
[ghstack-poisoned]
